Move superbase.ts into lib/api
