+++ src/hooks.client.ts
@@ -1,1 +1,1 @@
-import '$lib/supabase';
+import '$lib/api/supabase';
