+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { invalidateAll } from '$app/navigation';
-	import { supabase } from '$lib/supabase';
+	import { supabase } from '$lib/api/supabase';
 	import { onMount } from 'svelte';
 	import '../app.css';
